requirements.txt
# Google Sheets API
gspread>=6.0.0
google-auth>=2.25.2
google-auth-oauthlib>=1.1.0

# OpenAI Whisper and dependencies
openai-whisper>=20230918
torch>=2.2.0  # Updated for Python 3.12 compatibility
numpy>=1.26.0  # Compatible with Python 3.12
tqdm>=4.66.1
more-itertools>=10.1.0
transformers>=4.36.2
ffmpeg-python>=0.2.0

# Selenium and WebDriver
selenium>=4.15.2
webdriver-manager>=4.0.1
undetected-chromedriver>=3.5.5  # For enhanced stealth capabilities

# Anti-detection and proxy handling
fake-useragent>=1.4.0  # For generating realistic user agents
requests>=2.31.0
requests[socks]>=2.31.0  # For SOCKS proxy support

# Anthropic API for Claude
anthropic>=0.5.0

# Utilities
python-dotenv>=1.0.0
yt-dlp>=2023.10.13

# Performance and concurrency
aiohttp>=3.9.1  # Compatible with Python 3.12

# Note: futures is included in Python standard library as concurrent.futures
